+++ src/components/Trade/hooks/useSwapper/useSwapper.tsx
@@ -1,5 +1,4 @@
 import { type Asset } from '@shapeshiftoss/asset-service'
-import { fromAccountId } from '@shapeshiftoss/caip'
 import type { UtxoBaseAdapter } from '@shapeshiftoss/chain-adapters'
 import {
   type Swapper,
@@ -11,12 +10,13 @@ import type { KnownChainIds } from '@shapeshiftoss/types'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useFormContext, useWatch } from 'react-hook-form'
 import { useSelector } from 'react-redux'
+import { useReceiveAddress } from 'components/Trade/hooks/useReceiveAddress'
 import { getSwapperManager } from 'components/Trade/hooks/useSwapper/swapperManager'
 import {
   isSupportedNonUtxoSwappingChain,
   isSupportedUtxoSwappingChain,
 } from 'components/Trade/hooks/useSwapper/typeGuards'
-import { filterAssetsByIds, getReceiveAddress } from 'components/Trade/hooks/useSwapper/utils'
+import { filterAssetsByIds } from 'components/Trade/hooks/useSwapper/utils'
 import { useTradeQuoteService } from 'components/Trade/hooks/useTradeQuoteService'
 import type { TS } from 'components/Trade/types'
 import { type BuildTradeInputCommonArgs } from 'components/Trade/types'
@@ -24,7 +24,6 @@ import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingl
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { toBaseUnit } from 'lib/math'
 import { selectAssetIds, selectFeeAssetByChainId } from 'state/slices/assetsSlice/selectors'
-import { isUtxoAccountId } from 'state/slices/portfolioSlice/utils'
 import { selectFeatureFlags } from 'state/slices/preferencesSlice/selectors'
 import {
   selectBIP44ParamsByAccountId,
@@ -44,7 +43,6 @@ export const useSwapper = () => {
   const buyTradeAsset = useWatch({ control, name: 'buyTradeAsset' })
   const quote = useWatch({ control, name: 'quote' })
   const sellAssetAccountId = useWatch({ control, name: 'sellAssetAccountId' })
-  const buyAssetAccountId = useWatch({ control, name: 'buyAssetAccountId' })
   const isSendMax = useWatch({ control, name: 'isSendMax' })
   const isExactAllowance = useWatch({ control, name: 'isExactAllowance' })
   const slippage = useWatch({ control, name: 'slippage' })
@@ -59,11 +57,9 @@ export const useSwapper = () => {
   // Hooks
   const [swapperManager, setSwapperManager] = useState<SwapperManager>(() => new SwapperManager())
   const [bestTradeSwapper, setBestTradeSwapper] = useState<Swapper<KnownChainIds>>()
-  const {
-    state: { wallet },
-  } = useWallet()
-  const [receiveAddress, setReceiveAddress] = useState<string | null>()
+  const wallet = useWallet().state.wallet
   const { tradeQuoteArgs } = useTradeQuoteService()
+  const { receiveAddress } = useReceiveAddress()
 
   // Selectors
   const flags = useSelector(selectFeatureFlags)
@@ -83,6 +79,7 @@ export const useSwapper = () => {
     [assetIds, swapperManager],
   )
 
+  // Selectors
   const sellAssetAccountIds = useAppSelector(state =>
     selectPortfolioAccountIdsByAssetId(state, { assetId: sellAsset?.assetId ?? '' }),
   )
@@ -99,17 +96,6 @@ export const useSwapper = () => {
     selectBIP44ParamsByAccountId(state, sellAccountFilter),
   )
 
-  const buyAssetAccountIds = useAppSelector(state =>
-    selectPortfolioAccountIdsByAssetId(state, { assetId: buyAsset?.assetId ?? '' }),
-  )
-  const buyAccountFilter = useMemo(
-    () => ({ accountId: buyAssetAccountId ?? buyAssetAccountIds[0] }),
-    [buyAssetAccountId, buyAssetAccountIds],
-  )
-  const buyAccountMetadata = useAppSelector(state =>
-    selectPortfolioAccountMetadataByAccountId(state, buyAccountFilter),
-  )
-
   const swapperSupportsCrossAccountTrade = useMemo(() => {
     if (!bestTradeSwapper) return false
     switch (bestTradeSwapper.name) {
@@ -123,31 +109,6 @@ export const useSwapper = () => {
         return false
     }
   }, [bestTradeSwapper])
-
-  const getReceiveAddressFromBuyAsset = useCallback(
-    async (buyAsset: Asset) => {
-      if (!buyAssetAccountId) return
-      if (!buyAccountMetadata) return
-      const { accountType, bip44Params } = buyAccountMetadata
-      if (isUtxoAccountId(buyAssetAccountId) && !accountType)
-        throw new Error(`Missing accountType for UTXO account ${buyAssetAccountId}`)
-      const buyAssetChainId = buyAsset.chainId
-      const buyAssetAccountChainId = fromAccountId(buyAssetAccountId).chainId
-      /**
-       * do NOT remove
-       * super dangerous - don't use the wrong bip44 params to generate receive addresses
-       */
-      if (buyAssetChainId !== buyAssetAccountChainId) return
-      const receiveAddress = await getReceiveAddress({
-        asset: buyAsset,
-        wallet,
-        bip44Params,
-        accountType,
-      })
-      return receiveAddress
-    },
-    [buyAssetAccountId, buyAccountMetadata, wallet],
-  )
 
   const getSupportedBuyAssetsFromSellAsset = useCallback(
     (assets: Asset[]): Asset[] | undefined => {
@@ -271,28 +232,12 @@ export const useSwapper = () => {
     })()
   }, [flags])
 
-  // Set the receiveAddress when the buy asset changes
-  useEffect(() => {
-    const buyAsset = buyTradeAsset?.asset
-    if (!buyAsset) return
-    ;(async () => {
-      try {
-        const receiveAddress = await getReceiveAddressFromBuyAsset(buyAsset)
-        setReceiveAddress(receiveAddress)
-      } catch (e) {
-        setReceiveAddress(null)
-      }
-    })()
-  }, [buyTradeAsset?.asset, getReceiveAddressFromBuyAsset])
-
   return {
     getSupportedSellableAssets,
     getSupportedBuyAssetsFromSellAsset,
     swapperManager,
     checkApprovalNeeded,
     bestTradeSwapper,
-    receiveAddress,
-    getReceiveAddressFromBuyAsset,
     getTrade,
     swapperSupportsCrossAccountTrade,
     approve,
